requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic
oauthlib
requests
jinja2
python-jose